pre-finalizado

+++ routes/web.php
@@ -21,43 +21,43 @@ Route::get('/', function () {
 
 Route::group(['prefix'=>'Noticia','middleware'=>'auth'],function(){
 
-	Route::get('crear',[controlador\NoticiaController::class,'create'])->name('noticia.crear');
-	Route::post('crear',[controlador\NoticiaController::class,'store'])->name('noticia.guardar');
-	Route::get('listar',[controlador\NoticiaController::class,'list'])->name('noticias.listar');
-	Route::get('ver/{id}',[controlador\NoticiaController::class,'show'])->name('ver.noticia');
-	Route::get('ver/edit/{id}',[controlador\NoticiaController::class,'edit'])->name('ver.edit');
-	Route::put('actualizar/{id}',[controlador\NoticiaController::class,'update'])->name('noticia.editar');
-	Route::put('actualizar/foto/{id}',[controlador\NoticiaController::class,'updatePhoto'])->name('noticia.editarFoto');
-	Route::delete('eliminar/{id}',[controlador\NoticiaController::class,'destroy'])->name('noticia.eliminar');
+	Route::get('crear',[controlador\NoticiaController::class,'create'])->name('noticia.crear')->middleware(['permission:crear noticia']);
+	Route::post('crear',[controlador\NoticiaController::class,'store'])->name('noticia.guardar')->middleware(['permission:crear noticia']);
+	Route::get('listar',[controlador\NoticiaController::class,'list'])->name('noticias.listar')->middleware(['permission:listar noticias']);
+	Route::get('ver/{id}',[controlador\NoticiaController::class,'show'])->name('ver.noticia')->middleware(['permission:ver noticia']);
+	Route::get('ver/edit/{id}',[controlador\NoticiaController::class,'edit'])->name('ver.edit')->middleware(['permission:modificar noticia']);
+	Route::put('actualizar/{id}',[controlador\NoticiaController::class,'update'])->name('noticia.editar')->middleware(['permission:modificar noticia']);
+	Route::put('actualizar/foto/{id}',[controlador\NoticiaController::class,'updatePhoto'])->name('noticia.editarFoto')->middleware(['permission:modificar noticia']);
+	Route::delete('eliminar/{id}',[controlador\NoticiaController::class,'destroy'])->name('noticia.eliminar')->middleware(['permission:eliminar noticia']);
 });
 
 /*Route admin*/
 Route::group(['prefix'=>'Admin','middleware'=>'auth'],function(){
 
 	/*rol*/
-	Route::get('crear/rol',[controlador\RolController::class,'create'])->name('crear.rol');
-	Route::get('rol/crear',[controlador\RolController::class,'index'])->name('insertar.rol');
-	Route::post('crear',[controlador\RolController::class,'store'])->name('guardar.rol');
-	Route::get('eliminar/rol/{id}',[controlador\RolController::class,'destroy'])->name('eliminar.rol');
-	Route::get('editar/rol/{role}',[controlador\RolController::class,'show'])->name('editar.rol');
-	Route::put('editar/rol/{id}',[controlador\RolController::class,'update'])->name('update.rol');
-	Route::get('ver/rol/{role}',[controlador\RolController::class,'ver'])->name('ver.rol');
+	Route::get('crear/rol',[controlador\RolController::class,'create'])->name('crear.rol')->middleware(['permission:listar roles']);
+	Route::get('rol/crear',[controlador\RolController::class,'index'])->name('insertar.rol')->middleware(['permission:crear rol']);
+	Route::post('crear',[controlador\RolController::class,'store'])->name('guardar.rol')->middleware(['permission:crear rol']);
+	Route::get('eliminar/rol/{id}',[controlador\RolController::class,'destroy'])->name('eliminar.rol')->middleware(['permission:eliminar rol']);
+	Route::get('editar/rol/{role}',[controlador\RolController::class,'show'])->name('editar.rol')->middleware(['permission:modificar rol']);
+	Route::put('editar/rol/{id}',[controlador\RolController::class,'update'])->name('update.rol')->middleware(['permission:modificar rol']);
+	Route::get('ver/rol/{role}',[controlador\RolController::class,'ver'])->name('ver.rol')->middleware(['permission:ver rol']);
 
 	/*end rol*/
 
 	/*permissions*/
-	Route::get('crear/permiso',[controlador\PermissionController::class,'create'])->name('crear.permisos');
-	Route::post('crear/permiso',[controlador\PermissionController::class,'store'])->name('guardar.permiso');
-	Route::get('eliminar/permiso/{id}',[controlador\PermissionController::class,'destroy'])->name('eliminar.permiso');
+	Route::get('crear/permiso',[controlador\PermissionController::class,'create'])->name('crear.permisos')->middleware(['permission:acceso a permisos']);
+	Route::post('crear/permiso',[controlador\PermissionController::class,'store'])->name('guardar.permiso')->middleware(['permission:acceso a permisos']);
+	Route::get('eliminar/permiso/{id}',[controlador\PermissionController::class,'destroy'])->name('eliminar.permiso')->middleware(['permission:acceso a permisos']);
 	/*end permissions*/
 
 	/*users*/
-	Route::get('crear/usuario',[controlador\UserController::class,'create'])->name('crear.user');
-	Route::post('crear/usuario',[controlador\UserController::class,'store'])->name('guardar.user');
-	Route::get('listar/usuarios',[controlador\UserController::class,'index'])->name('listar.users');
-	Route::get('eliminar/usuario/{user}',[controlador\UserController::class,'destroy'])->name('eliminar.user');
-	Route::get('editar/usuario/{user}',[controlador\UserController::class,'edit'])->name('editar.user');
-	Route::put('actualizar/usuario/{user}',[controlador\UserController::class,'update'])->name('actualizar.user');
+	Route::get('crear/usuario',[controlador\UserController::class,'create'])->name('crear.user')->middleware(['permission:crear usuario']);
+	Route::post('crear/usuario',[controlador\UserController::class,'store'])->name('guardar.user')->middleware(['permission:crear usuario']);
+	Route::get('listar/usuarios',[controlador\UserController::class,'index'])->name('listar.users')->middleware(['permission:listar usuarios']);
+	Route::get('eliminar/usuario/{user}',[controlador\UserController::class,'destroy'])->name('eliminar.user')->middleware(['permission:eliminar usuario']);
+	Route::get('editar/usuario/{user}',[controlador\UserController::class,'edit'])->name('editar.user')->middleware(['permission:modificar usuario']);
+	Route::put('actualizar/usuario/{user}',[controlador\UserController::class,'update'])->name('actualizar.user')->middleware(['permission:modificar usuario']);
 	/*end users*/
 
 });
